+++ src/pages/game/index.tsx
@@ -1,1 +1,2 @@
 export { Game } from "./game";
+export { Config } from "./config";