+++ src/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-10-26
+
+### Added
+
+- Added React Router.
+- Added routes in router file.
+- Added config component.
+
+### Changed
+
+- Render RouterProvider in main.tsx.
+
 ## [0.2.0] - 2025-10-26
 
 ### Added